Refactor StrainKmerDatabase and related scripts for improved clarity and functionality
- Removed redundant code and comments in convert_db_format.py, streamlining the conversion process from pickle to Parquet.
- Enhanced error handling and validation in StrainKmerDatabase class, ensuring robust loading and processing of k-mer data from Parquet files.
- Updated parameter_config.py to clarify argument descriptions and improve consistency in argument handling.
- Renamed variables and adjusted docstrings for better readability and understanding of the code's purpose.
- Ensured that all paths are handled using pathlib for better cross-platform compatibility.
