+++ src/strainr/parameter_config.py
@@ -1,6 +1,5 @@
 import argparse
 import pathlib
-# from typing import List # List is not used in type hints in this file
 
 
 def process_arguments() -> argparse.Namespace:
@@ -16,29 +15,29 @@ def process_arguments() -> argparse.Namespace:
                             as attributes.
     """
 
-    parser: argparse.ArgumentParser = argparse.ArgumentParser(  # Renamed for clarity
-        description="Strainr: A tool for strain analysis using k-mer based methods."  # More descriptive
+    parser: argparse.ArgumentParser = argparse.ArgumentParser(
+        description="Strainr: A tool for strain analysis using k-mer based methods."
     )
     parser.add_argument(
         "input",
-        help="One or more forward/unpaired FASTQ input file(s).",  # Clarified help
+        help="One or more forward/unpaired FASTQ input file(s).",
         nargs="+",
-        type=pathlib.Path,  # Changed to pathlib.Path
+        type=pathlib.Path,
     )
     parser.add_argument(
         "-r",
         "--reverse",
-        help="Optional: One or more reverse FASTQ input file(s), corresponding to 'input'. (Feature: todo)",  # Clarified help
-        nargs="+",  # Should match number of input files if provided, or be single for all
-        type=pathlib.Path,  # Changed to pathlib.Path
-        default=[],  # Provide a default empty list
+        help="Optional: One or more reverse FASTQ input file(s), corresponding to 'input'. (Feature: todo)",
+        nargs="+",
+        type=pathlib.Path,
+        default=[],
     )
     parser.add_argument(
         "-d",
         "--db",
-        help="Path to the StrainKmerDatabase file (Parquet format).",  # Clarified help
-        type=pathlib.Path,  # Changed to pathlib.Path
-        required=True,  # Assuming database is essential for most operations
+        help="Path to the KmerStrainDatabase file (Parquet format).",  # Corrected
+        type=pathlib.Path,
+        required=True,
     )
     parser.add_argument(
         "-p",
@@ -47,7 +46,7 @@ def process_arguments() -> argparse.Namespace:
         default=4,
         help="Number of cores to use (default: 4)",
     )
-    parser.add_argument(  # Standardized to parser
+    parser.add_argument(
         "-o",
         "--out",
         type=pathlib.Path,
@@ -58,7 +57,7 @@ def process_arguments() -> argparse.Namespace:
     parser.add_argument(
         "-m",
         "--mode",
-        help=" Selection mode for diambiguation ",
+        help="Selection mode for disambiguation",  # Corrected
         choices=[
             "random",
             "max",
@@ -68,25 +67,24 @@ def process_arguments() -> argparse.Namespace:
         type=str,
         default="max",
     )
-    parser.add_argument(  # Standardized to parser
+    parser.add_argument(
         "-a",
         "--thresh",
-        help="Abundance threshold for reporting strains (default: 0.001).",  # Improved help
+        help="Abundance threshold for reporting strains (default: 0.001).",
         type=float,
         default=0.001,
     )
-    parser.add_argument(  # Changed from args to parser
+    parser.add_argument(
         "--bin",
         action="store_true",
         required=False,
-        help=" Perform binning.  ",
+        help="Perform binning.",  # Corrected
     )
-    parser.add_argument(  # Standardized to parser
+    parser.add_argument(
         "--save-raw-hits",
         action="store_true",
-        required=False,  # Default is False if not specified
-        help="Save intermediate k-mer hit scores and final assignments as pickle files.",  # Clarified help
+        required=False,
+        help="Save intermediate k-mer hit scores and final assignments to output files (format may vary).",  # Corrected
     )
-    # Ensure all calls are to 'parser', not 'args'
     config_space = parser.parse_args()
     return config_space
